+++ src/handlers/handlerMapping.js
@@ -1,8 +1,10 @@
 import { monsterKillHandler } from './monster.handler.js';
-import { towerAddOnHandler } from './towerAddOn.handler.js';
-import { towerClearHandler } from './towerClear.handler.js';
-import { towerRemoveHandler } from './towerRemove.handler.js';
-import { checkTowerLocation, towerAddOnHandler, towerClearHandler } from './tower.handler.js';
+import {
+  checkTowerLocation,
+  towerAddOnHandler,
+  towerClearHandler,
+  towerRemoveHandler,
+} from './tower.handler.js';
 import { gameStart, gameEnd } from './game.handler.js';
 
 const handlerMappings = {
